make client commons file
